refactored to a global fixture, and the reviewer's ask is satisfied.
